height slider logic complete
