+++ lib/logic/height/height_cubit.dart
@@ -6,5 +6,5 @@ class HeightCubit extends Cubit<HeightState> {
   HeightCubit() : super(HeightInitial());
   // void heightIncrement()=>emit(HeightState(height: state.height+1));
   // void heightDecrement()=>emit(HeightState(height: state.height-1));
-  void updateVal(double newValue)=>emit(HeightState(height: state.height=newValue.round()));
+  void updateVal(double newValue)=>emit(HeightState(height: state.height=newValue.toInt()));
 }
